requirements.txt
python-vlc==3.0.18121
schedule==1.1.0
playsound==1.3.0
